FloatingPoint: Removed two redundant type cast.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fix_floatingpoint_calculator@46255 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/floatingpoint/CMyFloat.java
@@ -109,7 +109,7 @@ public class CMyFloat extends CFloat {
   }
 
   private int calculateExpWidth(BinaryMathContext pFormat) {
-    return (int) Math.ceil(lb((double) (2 * pFormat.maxExponent + 1)));
+    return (int) Math.ceil(lb(2 * pFormat.maxExponent + 1));
   }
 
   private MyFloat parseFloat(String repr, BinaryMathContext pFormat) {
